requirements.txt
python-dotenv
Flask
aiogram==2.9
flask-sqlalchemy
flask-login
redis
aioredis
gino
